Fixed js components and added a method to copy static files

+++ copyStaticAssets.js
@@ -2,17 +2,20 @@ const shell = require("shelljs");
 const fs = require("fs");
 const path = require("path");
 
-console.log("Copy static assets...");
+console.log(" [assets] Copy static assets...");
 
 shell.cd("src");
 
 shell.exec(`find -name "*.html"     | xargs cp --parents -t ../dist`);
+console.log(" [assets] HTML files... Done");
 shell.exec(`find -name "*.css"      | xargs cp --parents -t ../dist`);
-shell.exec(`find -name "*.asset.js" | xargs cp --parents -t ../dist`);
+console.log(" [assets] CSS files... Done");
+shell.exec(`find -name "*.js"      | xargs cp --parents -t ../dist`);
+console.log(" [assets] Javascript files... Done");
+// shell.exec(`find -name "*.asset.js" | xargs cp --parents -t ../dist`);
 
-console.log("Done.\n");
-
-console.log("\n [browserify] Bundling all the UI scripts...");
+console.log(" [assets] Done.");
+console.log("\n [bundle] Bundling all the UI scripts...");
 
 // shell.exec(`find -name "dist/ui/*.js" | xargs browserify`);
 
@@ -45,19 +48,42 @@ walk(`${__dirname}/dist/ui/`, function (err, results) {
 
     for (let i = 0; i < results.length; i++) {
         let f = results[i];
-        if (f.endsWith(".js") && !f.endsWith(".asset.js")) {
+        if (f.endsWith(".js") && !f.endsWith(".asset.js") && !f.endsWith(".bundle.js")) {
             let output = f.replace(".js", ".bundle.js");
 
-            let parsed = path.parse(f);
-            console.log(" [compiling]", f, " -> ", output);
+            let shouldCompile = false;
+            try {
+                const outputFileMtime = fs.statSync(output).mtime;
+                const originalFileMtime = fs.statSync(f.replace(`${__dirname}/dist/ui/`, `${__dirname}/src/ui/`)).mtime;
 
-            shell.exec(`webpack ${f} --output-filename ${parsed.name}.bundle.js --output-path ${parsed.dir}`);
-            shell.rm(f);
-            shell.mv(output, f);
+                shouldCompile = originalFileMtime > outputFileMtime;
+            } catch (e) {
+                console.log(` [bundle] error: ${e}`)
+                shouldCompile = true;
+            }
 
-            console.log(" [bundle mv]", output, " -> ", f);
+            // shouldCompile = true; // for development or force compile
+
+            if (shouldCompile) {
+                let parsed = path.parse(f);
+                console.log(" [bundle]", f, " -> ", output);
+
+                let ex = shell.exec(`webpack ${f} --output-filename ${parsed.name}.bundle.js --target web --output-path ${parsed.dir}`);
+                if (ex.code != 0) { process.exit(1); }
+
+                console.log(" [bundle]", output, " -> ", f);
+
+                shell.rm(f);
+                shell.cp(output, f);
+            } else {
+                console.log(` [bundle] File '${f}' hasn't been changed, ignoring`);
+
+                shell.rm(f);
+                shell.cp(output, f);
+            }
         }
     }
+
+    console.log(" [bundle] Done.\n");
 });
 
-console.log(" [browserify] Done.\n");
